// source/apps/web-tenant/src/services/generated/models/createMenuItemDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * QR Ordering API
 * API Documentation for QR Ordering Platform
 * OpenAPI spec version: 1.0
 */

export interface CreateMenuItemDto {
  allergens?: string[];
  categoryId: string;
  description?: string;
  displayOrder?: number;
  imageUrl?: string;
  modifierGroupIds?: string[];
  name: string;
  price: number;
  tags?: string[];
}